+++ src/main.ts
@@ -12,6 +12,7 @@ import { FormView } from "./extensions/form-extension/obsidian";
 import javascriptExtension from "./extensions/javascript-extension";
 import minimalExtension from "./extensions/minimal-extension";
 import propertiesExtension from "./extensions/properties-extension";
+import snippetExtension from "./extensions/snippet-extension";
 import specialPropertiesExtension from "./extensions/special-properties-extension";
 import { getLogger } from "./logger";
 import { type ISettings, SettingTab } from "./setting-tab";
@@ -37,6 +38,7 @@ export default class PochoirPlugin extends Plugin {
         this.environment.extensions.use(formExtension());
         this.environment.extensions.use(commandExtension());
         this.environment.extensions.use(javascriptExtension());
+        this.environment.extensions.use(snippetExtension());
 
         this.app.workspace.onLayoutReady(() => {
             this.register(this.environment.enable());